==> llms/preconsultation/notebooks/1_Generate_train_val_test.ipynb <==
import os
import pandas as pd
import numpy as np
import datasets
from datasets import load_dataset

import torch
from collections import Counter
from transformers import AutoTokenizer
from transformers import AutoModelForSequenceClassification
from transformers import TrainingArguments, Trainer
from transformers import DataCollatorWithPadding
from sklearn import metrics
from sklearn.model_selection import train_test_split
from imblearn.over_sampling import RandomOverSampler

# Load the labeled dataset
keys = ['皮肤科', '胸外科', '妇科', '神经外科', '泌尿外科']

df1 = pd.read_excel("..\data\label_final_sentence_level.xlsx")
df1

label_cols = [ '信息支持1-病情描述/分析/诊断',
              '信息支持2-科普读物/其他病人经历',
     '信息支持3-线上/门诊或住院流程/时间/费用',
     '信息支持4-治疗建议',
             '信息支持',
             '情感支持'] 
stats = []
for i, key in enumerate(keys):
    data = df1[df1['key']==key]
    for col in label_cols:
        stats.append([key, col.split('-')[0], data[col].value_counts().values])
for col in label_cols:
    stats.append(['All', col.split('-')[0], df1[col].value_counts().values])

stats = pd.pivot(pd.DataFrame(stats), index=0, columns=1, values=2)
stats

label_cols = [ '信息支持1-病情描述/分析/诊断',
              '信息支持2-科普读物/其他病人经历',
     '信息支持3-线上/门诊或住院流程/时间/费用',
     '信息支持4-治疗建议',
             '信息支持',
             '情感支持']  

# Train model
text_col = '医生回复'
for label_col in label_cols:
    print(label_col)
    train_path = os.path.join('dataset_622', label_col[:5]+'_train.csv')
    val_path = os.path.join('dataset_622', label_col[:5]+'_val.csv')
    test_path = os.path.join('dataset_622', label_col[:5]+'_test.csv')
    data = df1
    columns = [text_col,label_col]
    print(columns)
    df = data[columns].copy()
    df.rename(columns={text_col: 'text'}, inplace=True)
    df.rename(columns={label_col: 'label'}, inplace=True)
    #split train and test
    df_train, df_test = train_test_split(df, test_size=0.40, random_state=42)
    #split validation and test
    df_val, df_test = train_test_split(df_test, test_size=0.50, random_state=42)
    df_test = df_test.reset_index(drop=True)
    df_val = df_val.reset_index(drop=True)
    #save val and test dataset
    df_val.to_csv(val_path, index=False)
    df_test.to_csv(test_path, index=False)
 
    #oversample the train
    label = 'label'
    df_train = df_train.reset_index(drop=True)
    X = df_train.drop(columns=[label])
    y = df_train[label]
    oversample = RandomOverSampler(random_state=42, sampling_strategy='minority')
    X_over, y_over = oversample.fit_resample(X, y)
    print(Counter(y_over))
    df_train_over = pd.DataFrame(X_over, columns=X.columns)
    df_train_over[label] = y_over
    df_train_over.to_csv(train_path, index=False)
