==> llms/preconsultation/notebooks/4_Text_split_sentences.ipynb <==
import pandas as pd
import re

df = pd.read_csv('hdf_service_20230117_sentences_with labels.csv', sep='\t', encoding='utf_8_sig', low_memory=True)
df

df = pd.read_csv('hdf_service_20230117.csv', sep='\t', encoding='utf_8_sig', low_memory=True)
df

#(705999, 52)
df[df['shanghai']==1].shape

df_shanghai = df[df['shanghai']==1]
df_shanghai

columns = ['service_id','all_doctor_text']
df1 = df_shanghai[columns].copy()
df1

# 使用逗号、句号、分号、感叹号和问号分割文本
def split_sentences(text):
    punctuation = "。；！？,.!?"
    sentences = []
    sentence = ""
    for char in text:
        if char in punctuation:
            if sentence:
                sentences.append(sentence.strip())
                sentence = ""
        else:
            sentence += char
    if sentence:
        sentences.append(sentence.strip())
    return sentences

# 将 "all_doctor_text" 列中的所有元素转换为字符串
df1['all_doctor_text'] = df1['all_doctor_text'].astype(str)
# 对 "all_doctor_text" 列应用分割函数
df1['all_doctor_text'] = df1['all_doctor_text'].apply(split_sentences)
# 将每个句子拆分成新的行
new_rows = []
for _, row in df1.iterrows():
    for sentence in row['all_doctor_text']:
        new_rows.append({"service_id": row["service_id"], "all_doctor_text": sentence})
# 创建新的 DataFrame
new_df1 = pd.DataFrame(new_rows)

# 打印生成的新 DataFrame
print(new_df1)

new_df1['service_id'].nunique()

df_cleaned = new_df1[new_df1['all_doctor_text'] != "['nan']"]
df_cleaned

df_cleaned['service_id'].nunique()
