==> llms/preconsultation/notebooks/5_Prediction.ipynb <==
import os
import pandas as pd
import numpy as np
import datasets
from datasets import load_dataset

import torch
from collections import Counter
from transformers import AutoTokenizer
from transformers import AutoModelForSequenceClassification
from transformers import TrainingArguments, Trainer
from transformers import DataCollatorWithPadding
from sklearn import metrics
from sklearn.model_selection import train_test_split
from imblearn.over_sampling import RandomOverSampler

curr_path = os.getcwd()
curr_path

df = pd.read_csv('hdf_service_20230117_sentences.csv', sep='\t', encoding='utf_8_sig')
#df.to_csv('test.csv',sep='\t', index=False)

df_test = pd.read_csv('test.csv', sep='\t', encoding='utf_8_sig')

prediction_folder = 'prediction'
if not os.path.exists(prediction_folder):
    os.makedirs(prediction_folder)

data_path = 'hdf_service_20230117_sentences.csv'
dataset = load_dataset('csv', delimiter='\t', data_files={'data': [data_path]})
print(dataset)
columns = dataset['data'].column_names
print(columns)

#label_col = '信息支持1-病情描述/分析/诊断'
#label_col = '信息支持2-科普读物/其他病人经历'
#label_col = '信息支持4-治疗建议'
#label_col = '信息支持'
#label_col= '情感支持'
#parameter = '_622_5e-5'
label_col = '信息支持3-线上/门诊或住院流程/时间/费用'
parameter = '_622_1e-4'
model_path = os.path.join(curr_path, 'Classifier', label_col[:5]+parameter +'_model')
print(model_path)
tokenizer = AutoTokenizer.from_pretrained(model_path)
def preprocess_function(examples):
    return tokenizer(examples['all_doctor_text'], truncation=True)
tokenized_dataset = dataset.map(preprocess_function, batched=True)
data_collator = DataCollatorWithPadding(tokenizer=tokenizer)
model = AutoModelForSequenceClassification.from_pretrained(model_path, num_labels=2) 
trainer = Trainer(
    model=model,
    tokenizer=tokenizer,
    data_collator=data_collator,
)
pred_dataset = tokenized_dataset["data"]
predictions = trainer.predict(pred_dataset)
preds = predictions.predictions.argmax(-1)
df = pd.read_csv('hdf_service_20230117_sentences.csv', sep='\t', encoding='utf_8_sig')
df[label_col]=preds
df[label_col].value_counts()
df.to_csv(os.path.join(curr_path, prediction_folder, 'prediction_'+label_col[:5]+'.csv'), index=False)
df

# df_test[label_col]=preds
# df_test[label_col].value_counts()
# #df_test.to_csv(os.path.join(curr_path, prediction_folder, 'prediction_'+label_col[:5]+'.csv'), index=False)
# df_test

df = pd.read_csv('/home/anqizhao/Anqi/Bert Classifier/hdf_service_20230117_sentences.csv', sep='\t', encoding='utf_8_sig')

df1=pd.read_csv('/home/anqizhao/Anqi/Bert Classifier/prediction/prediction_信息支持1.csv')
df2=pd.read_csv('/home/anqizhao/Anqi/Bert Classifier/prediction/prediction_信息支持2.csv')
df3=pd.read_csv('/home/anqizhao/Anqi/Bert Classifier/prediction/prediction_信息支持3.csv')
df4=pd.read_csv('/home/anqizhao/Anqi/Bert Classifier/prediction/prediction_信息支持4.csv')
df_info=pd.read_csv('/home/anqizhao/Anqi/Bert Classifier/prediction/prediction_信息支持.csv')
df_emo=pd.read_csv('/home/anqizhao/Anqi/Bert Classifier/prediction/prediction_情感支持.csv')

df1['id'] = df1.index + 1
df1

df2['id'] = df2.index + 1
df2

df3['id'] = df3.index + 1
df3

df4['id'] = df4.index + 1
df4

df_info['id'] = df_info.index + 1
df_info

df_emo['id'] = df_emo.index + 1
df_emo

merged_df = pd.merge(df2, df3, on =['service_id','all_doctor_text','id'],how='outer')
merged_df = pd.merge(merged_df, df4, on =['service_id','all_doctor_text','id'],how='outer')
merged_df = pd.merge(merged_df, df_info, on =['service_id','all_doctor_text','id'],how='outer')
merged_df = pd.merge(merged_df, df_emo, on =['service_id','all_doctor_text','id'],how='outer')
merged_df

merged_df.to_csv('/home/anqizhao/Anqi/Bert Classifier/hdf_service_20230117_sentences_with labels.csv', sep='\t', index=False)
