==> llms/preconsultation/notebooks/6_Naive_Bayes.ipynb <==
import pandas as pd

from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.naive_bayes import MultinomialNB

from sklearn.model_selection import cross_val_score
from sklearn.metrics import (accuracy_score, 
                             precision_score,
                             recall_score, 
                             f1_score)

from stopwordsiso import stopwords
import jieba

df_train = pd.read_csv("../data/informational_train.csv")
df_val = pd.read_csv("../data/informational_val.csv")
df_test = pd.read_csv("../data/informational_test.csv")

df_train

mnb = MultinomialNB()

X_train = df_train["text"]
y_train = df_train["label"]

X_val = df_val["text"]
y_val = df_val["label"]

X_test = df_test["text"]
y_test = df_test["label"]

stopwords_list = list(stopwords("zh"))

# Takes in a document, separates the words
def tokenize_zh(text):
    words = jieba.lcut(text)
    return words

vectorizer = TfidfVectorizer(tokenizer=tokenize_zh, stop_words=stopwords_list, min_df=30)

X_train = vectorizer.fit_transform(X_train).toarray()
X_val = vectorizer.transform(X_val).toarray()
X_test = vectorizer.transform(X_test).toarray()

len(X_train[0]), len(X_val[0]), len(X_test[0])

mnb.fit(X_train, y_train)

cv_scores = cross_val_score(mnb, X_val, y_val, cv = 10, scoring='accuracy')
cv_scores

predictions = mnb.predict(X_test)

accuracy = accuracy_score(y_test,predictions)
precision = precision_score(y_test, predictions)
recall = recall_score(y_test, predictions)
f1 = f1_score(y_test, predictions)

accuracy, precision, recall, f1
